Consolidate bar chart into single component and change how dimensions are handled
