--- context/app/static/js/components/Home/BarChartWrapper/index.js
@@ -1,3 +1,0 @@
-import BarChartWrapper from './BarChartWrapper';
-
-export default BarChartWrapper;
