--- context/app/static/js/components/Home/BarChartWrapper/style.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  grid-area: bar;
-`;
-
-export { Wrapper };
